+++ src/pages/Blog/Blog.module.scss
@@ -1,95 +1,86 @@
-@import url("https://fonts.googleapis.com/css2?family=Inter:wght@100..900&family=Jost:ital,wght@0,100..900;1,100..900&family=Quicksand:wght@300..700&display=swap");
-
+@import url("https://fonts.googleapis.com/css2?family=Montserrat:wght@700;800;900&display=swap");
 .blogpagecontainer {
-  font-family: "Montserrat", sans-serif;
-  color: white;
-  background: linear-gradient(rgba(0, 0, 0, 0.4), rgba(0, 0, 0, 0.4));
+  position: relative;
   min-height: 100vh;
-  padding-bottom: 3rem;
+  padding: 5rem 2rem 3rem;
+
+  &::before {
+    content: "";
+    position: absolute;
+    top: 0;
+    left: 0;
+    height: 100%;
+    width: 100%;
+    inset: 0;
+    background: url("../../assets/BlogBackground.png") no-repeat center;
+    background-size: cover;
+    filter: blur(1px) brightness(1.5);
+    z-index: 0;
+    // background-color: rgba(10, 10, 10, 0.3);
+    // background: linear-gradient(to bottom, rgba(0,0,0,0.1), rgba(0,0,0,0));
+    // backdrop-filter: blur(4px);
+
+    z-index: 0;
+  }
+
+  &::after {
+    content: "";
+    position: absolute;
+    inset: 0;
+    background-color: rgba(
+      78,
+      138,
+      190,
+      0.25
+    ); // ✅ dims it slightly without losing color
+    z-index: 1;
+  }
+
+  > * {
+    position: relative;
+    z-index: 2;
+  }
 }
 
 .pagetitle {
-  font-size: 3rem;
+  font-family: "Montserrat", "sans-serif";
+  font-size: 3.2rem;
   font-weight: 800;
   text-align: left;
-  padding: 6rem 3rem 0rem 5rem;
   color: white;
-
-  .underline {
-    display: block;
-    height: 6px;
-    width: 120px;
-    background: #4fc3f7;
-    margin-top: 0.5rem;
-  }
+  margin-left: 7rem;
+  margin-bottom: 2.5rem;
+  letter-spacing: 1px;
 }
 
+// .blogcardsgrid {
+//   display: grid;
+//   grid-template-columns: repeat(5, 1fr);
+//   gap: 2.2rem 1.5rem;
+//   justify-items: center;
+//   padding: 0 3rem;
+
+//   @media screen and (max-width: 1400px) {
+//     grid-template-columns: repeat(4, 1fr);
+//   }
+
+//   @media screen and (max-width: 1100px) {
+//     grid-template-columns: repeat(3, 1fr);
+//   }
+
+//   @media screen and (max-width: 768px) {
+//     grid-template-columns: repeat(2, 1fr);
+//   }
+
+//   @media screen and (max-width: 480px) {
+//     grid-template-columns: 1fr;
+//   }
 .blogcardsgrid {
-  display: grid;
-  grid-template-columns: repeat(5, 1fr); // 5 per row
-  gap: 2rem;
-  padding: 2rem 3rem;
-
-  @media screen and (max-width: 1200px) {
-    grid-template-columns: repeat(3, 1fr);
-  }
-
-  @media screen and (max-width: 768px) {
-    grid-template-columns: repeat(2, 1fr);
-  }
-
-  @media screen and (max-width: 480px) {
-    grid-template-columns: 1fr;
-  }
+  display: flex;
+  flex-wrap: wrap;
+  justify-content: center;
+  gap: 2.2rem 1.5rem;
+  padding: 0 3rem;
 }
 
-.blogcard {
-  background: white;
-  color: black;
-  max-height: 370px;
-  border-radius: 0px;
-  overflow: hidden;
-  box-shadow: 0 5px 15px rgba(0, 0, 0, 0.2);
-  text-align: center;
-
-  &:hover {
-    transform: scale(1.15);
-    transition: 1s all ease-in-out;
-  }
-
-  img {
-    width: 88%;
-    height: 180px;
-    padding-top: 16px;
-    object-fit: cover;
-    border-radius: 11px;
-    box-shadow: 0 5px 15px black;
-  }
-
-  .blogcardcontent {
-    font-family: "Jost", "sans-serif";
-    padding: 1rem;
-
-    h3 {
-      font-size: 1.3rem;
-      font-weight: 800;
-      margin: 0.5rem 0;
-    }
-
-    p {
-      font-size: 0.9rem;
-      margin-bottom: 1rem;
-      text-align: left;
-    }
-
-    a {
-      color: #000;
-      font-weight: 700;
-      text-decoration: none;
-
-      &:hover {
-        text-decoration: underline;
-      }
-    }
-  }
-}
+// }
